Implementazione testProprietarioDataImmatricolazioneSuperioreDi + correzione di errori nelle query

+++ src/it/prova/gestioneproprietari/dao/proprietario/ProprietarioDAO.java
@@ -5,5 +5,5 @@ import it.prova.gestioneproprietari.model.Proprietario;
 
 public interface ProprietarioDAO extends IBaseDAO<Proprietario> {
 	
-	public int contaQuantiConMacchinaDataImmatricolazioneDopo(int dataControllo) throws Exception;
+	public Long contaQuantiConMacchinaDataImmatricolazioneDopo(int dataControllo) throws Exception;
 }
